chore: 빌드 설정, firebase 초기화 구조 변경, service 파일을 firebase 모듈 측면에서 분리리 - manualChunks 의 ui 항목을 carousel 과 notification 으로 구체화 - firebase.ts 를 firebase 디렉토리의 모듈별 ts 파일로 분리   - app.ts, auth.ts, firestore.ts, storage.ts   - 관련 import 경로 모두 수정 - productService 내부의 image 관련 storage 로직을 분리리 -> imageService
