+++ src/firebase/app.ts
@@ -1,7 +1,4 @@
 import { initializeApp } from 'firebase/app';
-import { getAuth } from 'firebase/auth';
-import { getFirestore } from 'firebase/firestore';
-import { getStorage } from 'firebase/storage';
 import dotenv from 'dotenv';
 
 // Node.js 환경일 경우 dotenv 로드
@@ -27,7 +24,4 @@ const firebaseConfig = {
   measurementId: getEnv('VITE_measurementId'),
 };
 
-const app = initializeApp(firebaseConfig);
-export const auth = getAuth(app);
-export const db = getFirestore(app);
-export const storage = getStorage(app);
+export const app = initializeApp(firebaseConfig);
