+++ src/firebase/storage.ts
@@ -1,0 +1,4 @@
+import { getStorage } from 'firebase/storage';
+import { app } from './app';
+
+export const storage = getStorage(app);
