+++ src/services/productService.ts
@@ -1,4 +1,4 @@
-import { db, storage } from '@/firebase';
+import { buildFirestoreQuery, db } from '@/firebase/firestore';
 import { CartItem } from '@/hooks/useCartItems';
 import { UserInfo } from '@/hooks/useFirebaseAuth';
 import { OrderSchema, OrderStatus, ProductSchema } from '@/types/FirebaseType';
@@ -8,7 +8,6 @@ import {
   FetchInfiniteQueryResult,
   FetchQueryParams,
 } from '@/types/ReactQueryType';
-import { buildFirestoreQuery } from '@/utils/firebaseUtils';
 import { getKoreanIsoDatetime } from '@/utils/utils';
 import {
   collection,
@@ -24,23 +23,6 @@ import {
   startAt,
   where,
 } from 'firebase/firestore';
-import {
-  deleteObject,
-  getDownloadURL,
-  listAll,
-  ref,
-  uploadBytes,
-} from 'firebase/storage';
-
-// const updateBatch = async () => {
-//   let productsQuery = query(collection(db, 'product'));
-//   const productDocuments = await getDocs(productsQuery);
-//   const batch = writeBatch(db);
-//   productDocuments.docs.map((doc) =>
-//     batch.update(doc.ref, { productSalesrate: '0' }),
-//   );
-//   await batch.commit();
-// };
 
 /*
 ############################################################
@@ -118,13 +100,6 @@ export const getProductList = async (sellerId: string = 'all') => {
   } else return undefined;
 };
 
-export const uploadProductImage = async (path: string, imageFile: File) => {
-  const imageRef = ref(storage, path);
-  await uploadBytes(imageRef, imageFile);
-  const imageDownloadUrl = await getDownloadURL(imageRef);
-  return imageDownloadUrl;
-};
-
 export const updateProductData = async ({
   originalProductData,
   formData,
@@ -143,17 +118,6 @@ export const updateProductData = async ({
   };
 
   await setDoc(doc(db, 'product', originalProductData.id), documentData);
-};
-
-export const deleteProductImages = async (id: string) => {
-  // 판매 상품 경로의 이미지 폴더의 ref 를 불러옵니다.
-  const listRef = ref(storage, id);
-
-  // 폴더 내부의 파일 ref 를 모두 불러와서 각각에 대해 삭제 작업을 수행합니다.
-  const folderRef = await listAll(listRef);
-  for (const imageRef of folderRef.items) {
-    await deleteObject(imageRef);
-  }
 };
 
 export const deleteProductDocument = async (id: string) => {
